Actually compile again with latest version of v8
This doesn't run as expected just yet, it only compiles for now. An important
thing to note here is that you'll need to copy over 2 files from the v8 build directory
these being the natives_blob.bin and the snapshot_blob.bin. It requires these to actually
startup v8, but we're embedding it inside PHP-JS.